requirements.txt
newspaper
pika
python-dateutil
python-jsonrpc
pymongo
requests
redis
sklearn
